+++ library/Hasql/Transaction.hs
@@ -3,13 +3,14 @@
 module Hasql.Transaction
 (
   -- * Transaction monad
+  TransactionT,
   Transaction,
   condemn,
   abandon,
-  sql,
-  statement,
+  MonadSession (..),
 )
 where
 
+import Hasql.Session
 import Hasql.Transaction.Private.Transaction
 import Hasql.Transaction.Private.Model