+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 1.0.1.2
 
-- Use `MonadMask` for opening and closing the transaction
-- Add `unsafeRunIO`
-- Add `abandon`
+- Use `MonadMask` for opening and closing the transaction.
+- Make `Transaction` a transformer.
+- Add `unsafeLift` and `unsafeRunIO`.
+- Add `abandon`.
+- Make use of the `MonadSession` typeclass for `sql` and `statement`.